Compatibility with Python 3.4 and Python 2.7.6.
I had to fix some unit-test and add some encoding when reading files (because lxml can not create an XML document with encoding header from an unicode string).
Documentation update, with a small part for "contributing", and "compatibility".
Add a requirements.txt, because it's better for dev.
Bump setup.py and doc/conf.py version (0.2.x to 0.3.x).