Allow lxml 3.5 in setup.py and tox
After a short discussion on IRC about "is XML a terrible mistake?" it
appears that there is a new version of lxml 3.5 that is compatible with
this project.

In fact, I'm quite sure any version of lxml < 4.x will work fine.

Also, tox results show me that Python 2.7 with any version of lxml
works fine too. What a wonderfull world?

Thank you @mrjmad for the request! ;-)